+++ oap-server/analyzer/agent-analyzer/src/main/java/org/apache/skywalking/oap/server/analyzer/provider/trace/parser/listener/SourceBuilder.java
@@ -88,6 +88,9 @@ class SourceBuilder {
     private int responseCode;
     @Getter
     @Setter
+    private String responseCodeStr;
+    @Getter
+    @Setter
     private RequestType type;
     @Getter
     @Setter
@@ -119,6 +122,7 @@ class SourceBuilder {
         all.setLatency(latency);
         all.setStatus(status);
         all.setResponseCode(responseCode);
+        all.setResponseCodeStr(responseCodeStr);
         all.setType(type);
         all.setTimeBucket(timeBucket);
         all.setTags(tags);
@@ -137,6 +141,7 @@ class SourceBuilder {
         service.setLatency(latency);
         service.setStatus(status);
         service.setResponseCode(responseCode);
+        service.setResponseCodeStr(responseCodeStr);
         service.setType(type);
         service.setTags(tags);
         service.setTimeBucket(timeBucket);
@@ -159,6 +164,7 @@ class SourceBuilder {
         serviceRelation.setLatency(latency);
         serviceRelation.setStatus(status);
         serviceRelation.setResponseCode(responseCode);
+        serviceRelation.setResponseCodeStr(responseCodeStr);
         serviceRelation.setType(type);
         serviceRelation.setDetectPoint(detectPoint);
         serviceRelation.setTimeBucket(timeBucket);
@@ -178,6 +184,7 @@ class SourceBuilder {
         serviceInstance.setLatency(latency);
         serviceInstance.setStatus(status);
         serviceInstance.setResponseCode(responseCode);
+        serviceInstance.setResponseCodeStr(responseCodeStr);
         serviceInstance.setType(type);
         serviceInstance.setTags(tags);
         serviceInstance.setTimeBucket(timeBucket);
@@ -203,6 +210,7 @@ class SourceBuilder {
         serviceInstanceRelation.setLatency(latency);
         serviceInstanceRelation.setStatus(status);
         serviceInstanceRelation.setResponseCode(responseCode);
+        serviceInstanceRelation.setResponseCodeStr(responseCodeStr);
         serviceInstanceRelation.setType(type);
         serviceInstanceRelation.setDetectPoint(detectPoint);
         serviceInstanceRelation.setTimeBucket(timeBucket);
@@ -221,6 +229,7 @@ class SourceBuilder {
         endpoint.setLatency(latency);
         endpoint.setStatus(status);
         endpoint.setResponseCode(responseCode);
+        endpoint.setResponseCodeStr(responseCodeStr);
         endpoint.setType(type);
         endpoint.setTags(tags);
         endpoint.setTimeBucket(timeBucket);
@@ -252,6 +261,7 @@ class SourceBuilder {
         endpointRelation.setRpcLatency(latency);
         endpointRelation.setStatus(status);
         endpointRelation.setResponseCode(responseCode);
+        endpointRelation.setResponseCodeStr(responseCodeStr);
         endpointRelation.setType(type);
         endpointRelation.setDetectPoint(detectPoint);
         endpointRelation.setTimeBucket(timeBucket);