+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/source/All.java
@@ -56,6 +56,9 @@ public class All extends Source {
     private int responseCode;
     @Getter
     @Setter
+    private String responseCodeStr;
+    @Getter
+    @Setter
     private RequestType type;
     @Getter
     @Setter