+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/source/Endpoint.java
@@ -75,6 +75,9 @@ public class Endpoint extends Source {
     private int responseCode;
     @Getter
     @Setter
+    private String responseCodeStr;
+    @Getter
+    @Setter
     private RequestType type;
     @Getter
     @Setter